+++ apps/cowswap-frontend-e2e/src/e2e/swap.test.ts
@@ -65,12 +65,11 @@ describe('Swap (custom)', () => {
 
   it('can swap ETH for USDC', () => {
     cy.visit(`/#/${CHAIN_ID}/swap/ETH/${USDC}`, {
-      onBeforeLoad: async (win) => {
-        const address = await win.ethereum.signer.getAddress()
+      onBeforeLoad: (win) => {
         mockSendCall(win.ethereum, [
           handleNativeBalance(
             win.ethereum,
-            address,
+            win.ethereum.address,
             5n * 10n ** 18n, // 18 decimals
           ),
         ])
@@ -90,12 +89,11 @@ describe('Swap (custom)', () => {
   // ETH should be tradable but show Switch to Weth
   it('Swap ETH for USDC - shows optional Switch to WETH', () => {
     cy.visit(`/#/${CHAIN_ID}/swap/ETH/${USDC}`, {
-      onBeforeLoad: async (win) => {
-        const address = await win.ethereum.signer.getAddress()
+      onBeforeLoad: (win) => {
         mockSendCall(win.ethereum, [
           handleNativeBalance(
             win.ethereum,
-            address,
+            win.ethereum.address,
             5n * 10n ** 18n, // 18 decimals
           ),
         ])