+++ apps/cowswap-frontend/src/modules/trade/hooks/useWrapNativeFlow.ts
@@ -7,6 +7,7 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 import { WidgetHookEvents } from '@cowprotocol/widget-lib'
 
 import { Nullish } from 'types'
+import { usePublicClient, useWalletClient } from 'wagmi'
 
 import {
   wrapUnwrapCallback,
@@ -18,7 +19,7 @@ import { useTransactionAdder } from 'legacy/state/enhancedTransactions/hooks'
 
 import { buildTradeWidgetHookPayload, callWidgetHook } from 'modules/injectedWidget'
 
-import { useWethContract } from 'common/hooks/useContract'
+import { useWethContractData } from 'common/hooks/useContract'
 
 import { useDerivedTradeState } from './useDerivedTradeState'
 import { useWrapNativeScreenState } from './useWrapNativeScreenState'
@@ -66,10 +67,14 @@ function useWrapNativeCallback(inputAmount: Nullish<CurrencyAmount<Currency>>): 
 
 function useWrapNativeContext(amount: Nullish<CurrencyAmount<Currency>>): WrapUnwrapContext | null {
   const { account } = useWalletInfo()
-  const { contract: wethContract, chainId: wethChainId } = useWethContract()
+  const wethContract = useWethContractData()
+  const publicClient = usePublicClient()
+  const { data: walletClient } = useWalletClient()
   const addTransaction = useTransactionAdder()
   const [, setWrapNativeState] = useWrapNativeScreenState()
   const analytics = useCowAnalytics()
+
+  const wethChainId = wethContract.chainId
 
   return useMemo(() => {
     if (!wethContract || !amount || !account) {
@@ -80,6 +85,8 @@ function useWrapNativeContext(amount: Nullish<CurrencyAmount<Currency>>): WrapUn
       chainId: wethChainId,
       account,
       wethContract,
+      walletClient: walletClient ?? undefined,
+      publicClient: publicClient ?? undefined,
       amount,
       addTransaction,
       analytics,
@@ -90,5 +97,15 @@ function useWrapNativeContext(amount: Nullish<CurrencyAmount<Currency>>): WrapUn
         setWrapNativeState({ isOpen: true })
       },
     }
-  }, [wethChainId, wethContract, amount, addTransaction, setWrapNativeState, account, analytics])
+  }, [
+    wethChainId,
+    wethContract,
+    walletClient,
+    publicClient,
+    amount,
+    addTransaction,
+    setWrapNativeState,
+    account,
+    analytics,
+  ])
 }