+++ libs/tokens/src/state/tokens/userAddedTokensAtom.ts
@@ -6,7 +6,8 @@ import { getJotaiMergerStorage } from '@cowprotocol/core'
 import { mapSupportedNetworks } from '@cowprotocol/cow-sdk'
 import { Token } from '@cowprotocol/currency'
 import { PersistentStateByChain } from '@cowprotocol/types'
-import { getAddress } from '@ethersproject/address'
+
+import { getAddress } from 'viem'
 
 import { TokensMap } from '../../types'
 import { environmentAtom } from '../environmentAtom'
@@ -55,7 +56,7 @@ export const removeUserTokensAtom = atom(null, (get, set, tokens: string[]) => {
     delete stateCopy[token]
     delete stateCopy[token.toLowerCase()]
     try {
-      delete stateCopy[getAddress(token)]
+      delete stateCopy[getAddress(token as `0x${string}`)]
     } catch {}
   })