+++ libs/hook-dapp-lib/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [2.0.13](https://github.com/cowprotocol/cowswap/compare/hook-dapp-lib-v2.0.12...hook-dapp-lib-v2.0.13) (2026-04-10)
+
+
+### 🐛 Bug Fixes
+
+* **widget:** fix iframe issues ([#7298](https://github.com/cowprotocol/cowswap/issues/7298)) ([3f7094d](https://github.com/cowprotocol/cowswap/commit/3f7094df6389c01bdda515b645b58dcc0b292253))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/iframe-transport bumped to 2.0.13
+
 ## [2.0.12](https://github.com/cowprotocol/cowswap/compare/hook-dapp-lib-v2.0.11...hook-dapp-lib-v2.0.12) (2026-04-08)